Fix compilation warnings

+++ examples/net/zn_eval.c
@@ -21,7 +21,7 @@ char *value = "Eval from C!";
 void query_handler(zn_query_t *query, const void *arg) {
     const zn_string_t *res = zn_query_res_name(query);
     const zn_string_t *pred = zn_query_predicate(query);
-    printf(">> [Query handler] Handling '%.*s?%.*s'\n", res->len, res->val, pred->len, pred->val);
+    printf(">> [Query handler] Handling '%.*s?%.*s'\n", (int)res->len, res->val, (int)pred->len, pred->val);
     zn_send_reply(query, uri, (const unsigned char *)value, strlen(value));
 }
